CA-333610 avoid long running idle connection during migration
When migrating from a slave to another host, we were opening
a long running idle connection from the master to that slave.
This is because API requests are sent to the master and then
forwarded on to the appropriate slaves. That master <-> slave
connection was being held open for the duration of the migration,
with no traffic, so when migrations were taking long enough we
were hitting stunnel idle timeouts (in practice this was 12 hours,
because the default stunnel TIMEOUTidle value is 12 hours).

Internally, there were two ways to make requests to a host via the
XenAPI, with `Client.X.do_thing` and `Client.Async.X.do_thing`.
The former opens a connection which stays open until the request
has completed (obviously problematic in the case of migration, which
is slow), whereas the former spawns a new thread and sends an HTTP
response containing a task_ref promptly (the client is expected to
poll that task to wait until it has finished), meaning the connection
is also closed promptly.

To avoid long running idle connections during migrate_send we replace
`Client.VM.migrate_send` with `Client.InternalAsync.VM.migrate_send`, and
poll the resulting task_ref.

Ideally we would roll out similar changes to other API endpoints which
are slow, and also reduce the default TIMEOUTidle given to us by
stunnel.

Signed-off-by: lippirk <ben.anson@citrix.com>