CA-334811 assign xapi version automatically
When building with opam, we use `dune subst` to interpolate the version.

We cannot use `dune subst` when building the xapi rpms because we don't
have a git context at that point (`dune subst` uses `git describe --always
--dirty` internally). To get around this we expose a make target. For
completeness this will also work in a developer context, where git is
available.

Signed-off-by: lippirk <ben.anson@citrix.com>