CA-343230 assert (rather than assume) that update VBDs are attached
Ref counters are used to keep track of whether an update has a
mounted VBD or not. Unfortunately users have seen instances of
these counters becoming out of sync with the true state of the
VBDs.

One way of reproducing such a state is to upload an update
and then creating a VBD in dom0 corresponding to that update, i.e.

```
xe vbd-create vdi-uuid=$UPDATE_VDI_UUID vm-uuid=$DOM0_VM_UUID device=0
```

When trying to apply the update for the first time, xapi correctly
assumes the VBD you created is associated to the update you want
to apply, and tells you that it can't unplug it, because it was
never plugged in the first place. However, this puts the ref counters
in an invalid state, and when trying to apply the update a second time,
xapi looks at the ref counters and thinks that all of the VBDs have
been attached (when they haven't), and so continues the precheck
assuming all is well, but then when we try and fetch update.xml from
the VBD we inevitably fail.

Whilst it would be nice to fix the ref counters themselves this is
more difficult to implement than adding assertions to check whether
VBDs are attached or not.

Signed-off-by: lippirk <ben.anson@citrix.com>