REQ-403 copy_primary_host_certs API call
The motivation for this patch is that when joiners reboot, they may lose
out of cert updates (particularly when there are many parallel joins
occuring at the same time).

If a secondary host suspects that its local trusted certs are out of
date, then it can ask the primary host to install its trusted certs.

If cert checking is enabled, this will only work if the secondary still
has the primary's cert.

Signed-off-by: Ben Anson <ben.anson@citrix.com>